#18 Reorganize client scripts
