+++ observer/templates/index.html
@@ -17,6 +17,7 @@
 </head>
 
 <body>
+    <header></header>
     <div class="container">
         <div id="requests">
             <button type="button" id="ros2-topic-list">
